% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_map.R
\name{eq_map}
\alias{eq_map}
\title{Show Eqarthquoakes on a map with annotations}
\usage{
eq_map(data, annot_col)
}
\arguments{
\item{data}{is a filtered \code{data.frame} containing earthquake data
filtered to show quakes in an area of interest.}

\item{annot_col}{is a \code{string} which is column name carrying
annotations for marker pop-ups.  The annotation can be plain
text or HTML.}
}
\value{
\code{eq_map} returns a \code{leaflet} map object,
}
\description{
\code{eq_map} takes a \code{data.frame} and with NOAA
earthquake data and plots those on an interactive map, along with
annotations.  In addition to NOAA data, \code{eq_map} also takes a
parameter indicating which \code{data.frame} column to use for
marker, pop-up annotations.

\code{eq_map} expects to find epicenter lattiude, longitude and earthquake
magnitude in the \code{LATTITUDE}, \code{LONGITUDE} and \code{EQ_PRIMARY}
columns, respectively.
}
\examples{
library(quake)
library(readr)
library(dplyr)
library(lubridate)

data_file_name <- system.file("extdata", "earthquakes.tsv.gz", package = "quake")

read_delim(data_file_name, delim = "\\t") \%>\%
  eq_clean_data()                                       \%>\%
  filter(COUNTRY == "MEXICO" & year(date) >= 2000)      \%>\%
  mutate(popup_text = eq_create_label(.))               \%>\%
  eq_map(annot_col = "popup_text")
}
\seealso{
\link{leaflet}
}
