% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_create_label.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Create an HTML pop-up label showing location, magnitude and total_deaths}
\usage{
eq_create_label(data)
}
\arguments{
\item{data}{is a \code{data.frame} with \code{clean_location}, \code{EQ_PRIMARY}
and \code{TOTAL_DEATHS} columns.  The data in these columns is used to generate
the HTML encoded label strings.}
}
\value{
a vector of HTML encoded, label \code{string}s.
}
\description{
\code{eq_create_label} takes a NOAA \code{data.frame} with columns
\code{clean_location}, \code{EQ_PRIMARY} and \code{TOTAL_DEATHS} and produces
a vector of \code{string}s with HTML encoded labels showing the location,
magnitude and total, associated deaths caused the each quake.
}
\examples{
# Example with dplyr
library(quake)
library(readr)
library(dplyr)
library(lubridate)

data_file_name <- system.file("extdata", "earthquakes.tsv.gz", package = "quake")

read_delim(data_file_name, delim = "\\t") \%>\%
  eq_clean_data()                                       \%>\%
  filter(COUNTRY == "MEXICO" & year(date) >= 2000)      \%>\%
  mutate(popup_text = eq_create_label(.))               \%>\%
  eq_map(annot_col = "popup_text")
}
